recommend changes

+++ packages/protocol/contracts/L2/TaikoL2.sol
@@ -36,7 +36,7 @@ contract TaikoL2 is EssentialContract, TaikoL2Signer, IXchainSync {
      **********************/
 
     // TODO(david): need you to confirm this value.
-    uint64 public constant ANCHOR_GAS_COST = 47000;
+    uint32 public constant ANCHOR_GAS_COST = 47000;
 
     // Mapping from L2 block numbers to their block hashes.
     // All L2 block hashes will be saved in this mapping.
